dashboard/requirements.txt
streamlit
pandas
plotly-express
sqlalchemy
psycopg2-binary